feat: security audit completion — RBAC, mTLS, saga, outbox, payments, observability
Security:
- JWT hardening (iss/aud, methods, expiration, secret length check)
- Role-based access control (user/admin/service)
- mTLS between services
- Rate limiter with Redis-backed sliding window
- GraphQL introspection gated by env
- gRPC recovery interceptor

Reliability:
- Durable saga orchestrator with state machine and recovery
- Transactional outbox (order + catalog) with real Kafka publisher
- Inventory ledger for idempotent reservations
- Atomic payments with tx-manager
- Default timeouts on all downstream calls

Observability:
- OTLP exporter replacing deprecated jaeger
- Configurable logging (LOG_LEVEL, LOG_FORMAT)
- Health probes in all services
- pgx pool with OpenTelemetry and Prometheus metrics

Database:
- Money migration to int64 minor units
- CHECK constraints, indexes, triggers, NOT NULL timestamps
- Product stock split to inventory-only

Testing:
- Testcontainers for PG, Redis, ES, Kafka, ClickHouse
- gomock mocks and handler unit tests for all services
- Saga unit tests and E2E scenarios

CI:
- Actions pinned by SHA
- govulncheck, buf lint/breaking
- go.work module iteration for coverage

Refs: VERSION 0.2.0

+++ services/order-service/internal/saga/interfaces.go
@@ -8,6 +8,6 @@ type InventoryClient interface {
 }
 
 type PaymentClient interface {
-	ProcessPayment(ctx context.Context, orderID, userID string, amount float64) (string, error)
+	ProcessPayment(ctx context.Context, orderID, userID string, amount int64) (string, error)
 	Refund(ctx context.Context, paymentID string) error
 }
